refactor: Rename old files due to library name conflict
Change of jax and numpy submodules because they are conclicting with the libraries
