+++ fcmeans/_jax.py
@@ -1,4 +1,9 @@
-import numpy as np
+from jax import jit
+from jax import random
+from jax import numpy as np
+import time
+import logging
+logging.disable(logging.WARNING)
 
 
 class FCM:
@@ -66,14 +71,19 @@ class FCM:
 
     """
 
-    def __init__(self, n_clusters=10, max_iter=150, m=2, error=1e-5, random_state=42):
+    def __init__(
+        self, n_clusters=10, max_iter=150, m=2, error=1e-5, random_state=42
+    ):
         assert m > 1
         self.u, self.centers = None, None
         self.n_clusters = n_clusters
         self.max_iter = max_iter
         self.m = m
         self.error = error
-        self.rng = np.random.default_rng(random_state)
+        if not random_state:
+            self.key = random.PRNGKey(int(time.time()))
+        else:
+            self.key = random.PRNGKey(random_state)
 
     def fit(self, X):
         """Compute fuzzy C-means clustering.
@@ -84,7 +94,8 @@ class FCM:
             Training instances to cluster.
         """
         self.n_samples = X.shape[0]
-        self.u = self.rng.uniform(size=(self.n_samples, self.n_clusters))
+        self.u = random.uniform(key=self.key, shape=(
+            self.n_samples, self.n_clusters))
         self.u = self.u / np.tile(self.u.sum(axis=1)
                                   [np.newaxis].T, self.n_clusters)
         for iteration in range(self.max_iter):
@@ -131,11 +142,13 @@ class FCM:
         return self.__predict(X).argmax(axis=-1)
 
     @staticmethod
+    @jit
     def _dist(A, B):
         """Compute the euclidean distance two matrices"""
         return np.sqrt(np.einsum("ijk->ij", (A[:, None, :] - B) ** 2))
 
     @staticmethod
+    @jit
     def _next_centers(X, u, m):
         """Update cluster centers"""
         um = u ** m
@@ -148,7 +161,8 @@ class FCM:
             return np.sum(self.u ** 2) / self.n_samples
         else:
             raise ReferenceError(
-                "You need to train the model first. You can use `.fit()` method to this."
+                "You need to train the model first. You can use `.fit()` "
+                "method to this."
             )
 
     @property
@@ -157,5 +171,6 @@ class FCM:
             return -np.sum(self.u * np.log2(self.u)) / self.n_samples
         else:
             raise ReferenceError(
-                "You need to train the model first. You can use `.fit()` method to this."
+                "You need to train the model first. You can use `.fit()` "
+                "method to this."
             )
